requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
pandas==2.1.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
aiofiles==23.2.1
jinja2==3.1.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
pandas==2.1.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
aiofiles==23.2.1
jinja2==3.1.2
